fix: Move send_confirmation_email from webhook_handler to views
issue: The email did not match the total if a user redeemed their points. reason: Stripe webhook_handler gets the payment data when the checkout page is loaded and sent email with that data. However, a user can redeem their points on the checkout page after that data sent to Stripe